projects/backend/requirements.txt
# FastAPI backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
prisma>=0.13.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Authentication
better-auth>=0.0.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# AI/ML
litellm>=1.40.0
openai>=1.12.0
sentence-transformers>=2.3.0

# Vector store
qdrant-client>=1.7.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0